allow overwriting default variables
Most projects, especially outside GitHub, will not use the same URL for
their web page and repository, so $_giturl may be specified to give a
repository location differing from the PKGBUILD $url variable.

In a related manner, the repository may not necessarily named similar to
the $pkgname, especially when the package needs to check out multiple
repositories apart from the one the template works, so this can be
configured as $_gitname.