support arbitrary integrity checks
The static md5 variant is quite outdated. Borrowing functions from the
makepkg program itself, it is easy to automatically augment present sums
or, when none is yet present, act according to the preference expressed
in the makepkg.conf.

While bash's typeset builting supports indirect assignment through the
interpretation of variable names, it cannot utilize the += operator, so
an additional indirection step using a name variable is necessary.