keep pkgver function error free
Since commit 16f6aae33087ec04c1dc90aa84f7d3dde4333046 [1], which was
shipped out with Pacman version 5.2.0 the pkgver function is checked
like most other functions in the PKGBUILD: makepkg exits whenever any
command in it returns a non-zero exit code.

Reversing the condition checking for the return code of "git describe"
directs the non-zero variant to be overwritten by the exit code of the
revision counting command.

The line before is already guaranteed to succeed even when "git
describe" fails: Bash will only regard the exit code of the last command
in a pipe. (This is indeed specific to Bash.) As the Sed script is
reasonably safe, this should always be the case.

[1]: https://git.archlinux.org/pacman.git/commit/?id=16f6aae33087ec04c1dc90aa84f7d3dde4333046